requirements.txt
pytelegrambotapi~=4.23.0

environs~=11.0.0
psycopg2-binary~=2.9.9